Fixed a failure of power button press while entering G3 state.
If the power button is pressed while the S5 inactivity timer is
about to expire, EC need to give CPU a little time to start up
before changing the state from S5 to G3 (the hard off state).
Otherwise the system will not start up. This issue can be
reproduced on Rambi.

BUG=chrome-os-partner:42728, chrome-os-partner:42811
BRANCH=None
TEST=Implement an ec command to simulate power button press while
     S5 inactivity timer is about to expire, and then make sure
     that the patch did solve the issue.

Change-Id: I022e8e14fd41447898760a4d57a4702e2c00a0d5
Signed-off-by: Kaiyen Chang <kaiyen.chang@intel.com>
Signed-off-by: Icarus Sparry <icarus.w.sparry@intel.com>