Update patch set 1
Patch Set 1: Code-Review-1

Ah, I remember the problem now. This won't work as is because the BC1.2 interrupt is not configured, so nothing will wake the USB_CHG task. This is complicated by the fact that different revs of the board do this differently. Anyway, I'll leave this CL here in case you guys want to take it and figure out how to configure the BC1.2 interrupts.

Patch-set: 1
Label: Code-Review=-1