Update patch set 4
Patch Set 4:

I don't think we have an API for it. For keystrokes, the AP can simply ignore them at the UI or driver level. For powering down peripherals, the AP has no way to control it directly.

Maybe we can change the way things work in the future -- have a new host command where the AP can tell us that the lid is "closed enough" or "open enough", and the EC can take whatever actions are necessary. Can you open a crosbug.com about it?

Patch-set: 4
Label: Code-Review=0