lm4: Add debug output for overlapping HCs.
Currently, when a host command is received which would overlap with an
ongoing host command, we ignore it silently.  This commit simply logs a
line to the EC console stating that we are ignoring the overlapping host
command.

BRANCH=None
BUG=chrome-os-partner:23806
TEST=make -j buildall tests
TEST=Build, flash, and boot samus. Using ectool, read 64K from flash
while also querying the current EC switch positions.  Observe the log
message being printed to the EC console.

Change-Id: Ic0d249ccec2efb9600bcf8567392add1ee6295d9
Signed-off-by: Aseda Aboagye <aaboagye@google.com>