Update patch set 4
Patch Set 4:

Hi Marc,

It is really S5 state while waiting for the second S5 inactivity timeout event. So I think reuse S5 state is better than create another state for this.

Power button press event would not change the state directly, whether this patch is applied or not. The state will change from S5 to S5S3 only while EC detects that GPIO_PCH_SLP_S4_L is pulled high by CPU. So change to POWER_S5S0 (actually there is not a state named S5S0) when S5 inactivity timer expires is not correct.

To stay in S5 the state machine will handle all of the events just like what it would do while waiting for S5 inactivity timeout without this patch. If there are holes, I think they are already there without this patch. Can you just elaborate the holes you are worry about? Maybe we can try to solve it.

BTW, I'm not sure that if using want_g3_exit is right or not. Actually it is not in G3 state yet in this situation. My suggestion is, if there is no other functional concerns on this patch, we can submit it; then use want_g3_exit by another CL, and do more tests for it.

Thanks,
Kaiyen

Patch-set: 4