Update patch set 4
Patch Set 4:

Hi Zhuo-hao,

So you just let the system go to G3 and let it start up from G3 in this situation? I'm not sure this will work, because pwrbtn has been send to PCH, and if we let the system go to G3, RSMRST will be pulled low by EC in S5G3, that will make CPU go down, and EC will not send another pwrbtn to PCH again. (Maybe we should check want_g3_exit flag again in S5G3 to prevent it?) I will do some test to verify it, however, I still prefer to prevent it go to S5G3 from S5 in this situation, because the power state transition is simpler than this.

Patch-set: 4
Label: Code-Review=0