cortex-m: catch misconfigured interrupts early
The code in core/cortex-m/init.S limits the number of installed
vectors to CONFIG_IRQ_COUNT. But the DECLARE_IRQ macro installing
interrupt servicing routines does not care about this limitation. This
results in corrupted interrupt configuration, which is hard to debug.

This patch makes sure that there is a compilation error in case
DECLARE_IRQ is passed interrupt number which out of bounds.

A similar change needs to be introduced for cortex-m0.

BRANCH=none
BUG=chromium:518898
TEST=tried building cr50 with one of interrupt numbers exceeding
     CONFIG_IRQ_COUNT, observed a compilation error.

Change-Id: Ie7bc623da6bf7371579b2242064f81a83053df17
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>