strago: Disable TRACKPAD_PWREN during G3
TRACKPAD_PWREN should be off which is causing leakage
in 1P8A in Soc G3 case. The GPIO will be set to 0
during transition from S3 to S5 and will be set to 1
during transition from S5 to S3. This flow cannot be
done in S0-S3 transitions because TRACKPAD is a wake
source from S3.

BUG=None
BRANCH=strago-7287.B
TEST=Verify the power leakage in 1P8A rails in Soc G3 case

Change-Id: Ibcb903b9f47be0e4a086560a477499a7364c441d
Signed-off-by: Kumar, Gomathi <gomathi.kumar@intel.com>