celes: High EC power leakage fix
This patch fixes the EC high power consumption during S3 in BCRD2.
The GPIO201 was configured as OUTPUT and was showing leakage
current in S3 state. This GPIO is not used by EC at all. To fix
it configure it as INPUT to avoid leakage.

BRANCH=None
BUG=None
TEST=Build and boot the celes board. Measure EC and system
power during S3

Change-Id: I0fed64158ab9a92fd54459df0498928141efa038
Signed-off-by: Wonjoon Lee <woojoo.lee@samsung.com>