Strago: High EC power leakage fix
This patch fixes the EC high power consumption during S3 in BCRD2.
The GPIO201 was configured as OUTPUT and was showing leakage
current in S3 state. This GPIO is not used by EC at all. To fix
it configure it as INPUT to avoid leakage.

BRANCH=None
BUG=None
TEST=Build and boot the BCRD2 board. Measure EC and system
power during S3

Change-Id: Ibd47faef5d933d16a4a5f7b2c64a6fc3d4ad8529
Signed-off-by: Shobhit Srivastava <shobhit.srivastava@intel.com>