Kunimitsu: Fix for leakage in the Kepler
In girq19_interrupt handler setup_lpc function is called if the
MEC1322_LPC_BUS_MONITOR is set immediately. However on Kunimitsu,
due to the leakage in the Kepler, LPC is taking time to set the
MEC1322_LPC_BUS_MONITOR bit. As setup_lpc is not getting called
communication between the EC & Coreboot is not happening which
is preventing the board to boot to UI. Added delay by using a loop
(found by trial and error method * 2) to bypass this issue.

BUG=none
TEST=Kunimitsu device boots to UI.
BRANCH=none

Change-Id: Ie03f74f00ff8733b76597290242496c3d46e1a1b
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>