Kunimitsu: Fix for LPC init failure
In girq19_interrupt handler, setup_lpc function is called if the
MEC1322_LPC_BUS_MONITOR bit deasserted immediately. However on Kunimitsu
FAB3, due to the leakage in the Kepler, LPC is taking time to deassert
MEC1322_LPC_BUS_MONITOR bit. As setup_lpc is not getting called,
communication between the EC & Coreboot is not happening which is
preventing the board to boot to UI.
Waiting for MEC1322_LPC_BUS_MONITOR bit to deassert in a definite loop
to bypass this issue.

FAB4 has the hardware fix and this patch will be reverted for FAB4.

BUG=chrome-os-partner:43996
TEST=Kunimitsu FAB3 device boots to UI.
BRANCH=none

Change-Id: Ie03f74f00ff8733b76597290242496c3d46e1a1b
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>