ryu: Change charger termination current to 320mA.
Good rule of thumb for charger current is that it should be set
between C/20 & C/50 for the battery of the device.

For ryu its presently set at 64mA which may explain why we see
charger's status as charging even when battery is full.  In any case
its well below recommendation so lets change it to something within
that.  320 is just that number (multiple of 64).

Note, also changing bq2589x default termination to match its POR value
of 256 as no other board uses this charger.

Signed-off-by: Todd Broch <tbroch@chromium.org>

BRANCH=smaug
BUG=chrome-os-partner:42848
TEST=manual, read charger IC termination register and see termination
current set to 320mA

Change-Id: I60dbb9326c3abb8091fd9ab18eda08b9eabb197b