Update patch set 2
Patch Set 2:

FSP is forcing SLP_SUS# minimum assertion width to 4 seconds.  I'm not sure if this is necessary but if it is we need to account for it somehow (magically know the value?) in the EC.  If I disable the minimum assertion width for SLP_SUS# then it behaves consistently with this patch.

Patch-set: 2
Label: Code-Review=0