Fixed a failure of power button press while entering G3 state.
If the power button is pressed while S5 inactivity timer is about to
expire, EC needs to give CPU a little time to start up before changing
the state from S5 to G3 (the hard off state); otherwise the system will
not start up. This issue can be reproduced on Cyan.

BUG=chrome-os-partner:42994
BRANCH=None
TEST=Implement an ec command to simulate power button press while S5
     inactivity timer is about to expire, and then make sure that the
     patch did solve the issue.

Change-Id: I704aa6432f3faaa3c8a2e148d4072b4e78651ebf
Signed-off-by: Shamile Khan <shamile.khan@intel.com>