oak: add battery setting for rev3 & rev4
The battery type of rev3 is changed to 3S, add this setting to battery_info.
But in order to compatible to 2S battery, set the minimum voltage of
battery to 6V rather than 9V.
For rev4, it should always connect to 3S battery and doesn't necessary to
support 2S battery.

BRANCH=none
BUG=none
TEST=manual
1. connect battery of rev3.
2. check the battery info with EC command:
   > battery
3. confirm the voltage of battery is match to setting of battery_info.

Change-Id: I626ed150d76a1dfba7d1caf08c4f19882e9cb089
Signed-off-by: Ben Lok <ben.lok@mediatek.com>