i2c: Fix SCL unwedge logic
The current logic breaks out of the for() loop if SCL gets unwedged
(goes high), but still falls through to the "I2C unwedge failed,
SCL is being held low" case.  Fix this so that we only hit the
"SCL is being held low" case if SCL actually is stuck low.

BUG=none
BRANCH=link
TEST=manual

Change-Id: Ibe3f956db90bff88c8c1e380c5413f5a340cf268
Signed-off-by: Kevin Cernekee <cernekee@chromium.org>