common: make sure changing RO_MEM_OFF is accounted for in the image size
In some cases it is necessary to allocate room in the head of the RO
flash space, which is possible by setting CONFIG_RO_MEM_OFF. The
problem with this is that it is not accounted for when calculating the
actual RO image size.

This patch fixes the calculations.

BRANCH=none
BUG=chrome-os-partner:43025
TEST=with this and other changes the latest cr50 image gets signed and
     booted properly.

Change-Id: Icabf00c6b90657b0ae3370ddad3567ccedf3ff49
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>