Producer: Make a printf producer for printing to queues
This provides a set of printing routines (putc, puts, printf, and vprintf)
that drive a producer object that can be tied to a queue.  The result is a
clean way to print directly into a queue that can be connected to a
streaming device driver (like the USB stream or multi USART drivers). This
also provides a blocking version of the printf producer that will block up
to a configurable timeout waiting for space in the queue.

Signed-off-by: Anton Staaf <robotboy@chromium.org>

BRANCH=None
BUG=None
TEST=make buildall -j

Change-Id: Ib543080065448463ead8ede4f59b781a7bce09d7