Update patch set 4
Patch Set 4:

But I'm not sure if chipset_exit_hard_off() is always triggered by the tasks the priority of which is lower than the priority of chipset task on all of the platforms now and in the future. If chipset_exit_hard_off() is triggered by a task whose priority is higher than chipset task, the race condition will still have chance to occur. I still prefer to use CL:290280 to solve it, after all it has no any other concern.

Patch-set: 4