Kunimitsu: Workaround for PMIC emergency shutdown
When the DUT enters to SOC G3 or rebooted, SLP_SUS assertion turns off the
A-rails which is causing the PMIC emergency shutdown. As a workaround this
patch disables the power fault in the PMIC register.

This patch is applicable for FAB3 only and will be reverted when the
reboot logic in the hardware is fixed.

BUG=chrome-os-partner:44693
BRANCH=none
TEST=Manually tested on Kunimitsu FAB3.
     - "reboot" from the EC console command works
     - "ectool reboot_ec" from the Kernel terminal works
     - "shutdown -h now" command from the Kernel terminal puts the device
       in SoC G3 / PG3.
     - cold reset from the servo board works

Change-Id: Id5e091ace876d7655f64e61cca4a9f0303b69604
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>