Remove unnecessary want_g3_flag setting and resetting.
chipset_exit_hard_off() is always triggered by the tasks the priorities
of which are lower than the priority of chipset task. Therefore we don't
need to worry about that chipset_exit_hard_off() is invoked while the
power state is in POWER_S5G3, or is in POWER_S5 but the S5 inactivity
timer has expired. Removing unnecessary want_g3_flag setting and
resetting to let it be better structured.

BUG=None
BRANCH=None
TEST=Run consecutive reboot test 5000 times.

Change-Id: Id1fee6a6d5b6e3bd4fcafff795cc741243759a74
Signed-off-by: Kaiyen Chang <kaiyen.chang@intel.com>