Update patch set 4
Patch Set 4:

Actually the reason to set flag in chipset_exit_hard_off() for POWER_S5G3 is the same with it for POWER_S5. If we can make sure that chipset_exit_hard_off() is always triggered by the tasks the priorities of which are lower than the priority of chipset task on all of the platforms now and in the future, we can apply https://chromium-review.googlesource.com/296027, which removed all unnecesary want_g3_exit flag setting and unsetting. It also removed the flag resetting in POWER_S5 state, because the flag is only set in G3 state and it is always be reset if the system want to exit G3. I think to modify them in one patch is more consistent and reasonable, if we can make sure there is no potential risk about chipset_exit_hard_off() invoking.

Patch-set: 4