Update patch set 1
Patch Set 1:

We're wasting too much time on this. I'm going to go back and revert all derivative CLs, then apply https://chromiumreview.googlesource.com/#/c/290280/ to all relevant branches. That CL was reviewed in-depth and it's the CL we're going with. If anyone has improvements / changes then wait until that CL lands on TOT, push your changes for review, and convince the reviewers that your changes are worthwhile.

Sorry for waffling on this, and thank you all for your contributions to this project.

Patch-set: 1