usb_charger: allow charging from vbus with no type-C connection
Allow charging from VBUS supplier even if there is no type-C
connection by declaring a VBUS supplier as a dedicated charger
if VBUS has been connected for 500ms with no type-C connection.
This allows us to charge from non-compliant cables with no Rp
on the CC line but that still provide VBUS.

BUG=b:23509406
BRANCH=smaug
TEST=tested on samus and ryu with A to C cable with no Rp.
verified that we charger and charging icon turns on.

Change-Id: I4cd75d57bc2d6711299cd9dc26b324d28c1614c0
Signed-off-by: Alec Berg <alecaberg@chromium.org>