Update patch set 4
Patch Set 4:

(4 comments)

I see, I did not realize that the panel ID signals were coming from a connector to the panel itself. Looking at the schematic perhaps these should be called pwr_id or the schematic should change the net names to panel_id?

I don't see where these get pulled to high by default in the schematic, is this relying on default behavior of the EC GPIOs to pull high?

It seems reasonable to detect the no panel case with 0x7 and not just fall through leaving the power in the default state.

Patch-set: 4
Label: Code-Review=0