charge_manager: fix override command for off and dont charge
Fix the charge override command when given OVERRIDE_OFF or
OVERRIDE_DONT_CHARGE args in order to guarantee that we switch
ports to sourcing power when they are connected to dual-role
devices.

OVERRIDE_OFF: turns off override and asks for a power swap
to become a source on the previous override port.
OVERRIDE_DONT_CHARGE: disables charging and attempts to
become a power source on all ports.

BUG=none
BRANCH=smaug
TEST=tested on ryu with chargeoverride host command

Change-Id: I8e5265101706f19571a695f1fe82104165a1ca88
Signed-off-by: Alec Berg <alecaberg@chromium.org>