Update patch set 1
Patch Set 1:

I'm not sure this is the best way to do this. The code in the board directory would need to be copied to any other board with CONFIG_USB_PD_TCPM_VBUS, and I'm not sure the code added to usb_pd_tcpc doesn't conflict with the spec. My thought would be to put this into tcpm_init(), to ask the TCPC the state of VBUS after initializing the alert mask.

Patch-set: 1