Update patch set 1
Patch Set 1: Verified+1 Trybot-Ready+1

This is just the first pass at the driver. With this we can perform reading from the sensor. It was heavily leveraged from the existing KXCJ9.

The items that I have not verified are the interrupt functionality.

Also, depending on how the double tap/gesture stuff goes we can create an interface to the HW's double tap. 

Lastly, sample data with the lid sitting on my desk.
    [446.048490 event 0x80000000 Lid=-62  , 0    , -1018 ]
    [446.058621 event 0x80000000 Lid=-60  , -1   , -1015 ]
    [446.068949 event 0x80000000 Lid=-63  , 0    , -1018 ]
    [446.079102 event 0x80000000 Lid=-66  , 0    , -1027 ]

Patch-set: 1
Label: Trybot-Ready=+1
Label: Verified=+1