GLaDOS: Add lid accelerometer.
This commit adds the lid accelerometer to the list of motion sensors on
the board.  It currently wrapped behind an ifdef for
HAS_TASK_MOTIONSENSE due to space contraints.

BUG=chrome-os-partner:43494
BRANCH=None
TEST=Build GLaDOS EC with driver enabled and verify that valid
accelerometer data is read, and that range, resolution, and odr can all
be modified.
TEST=Verified that signs of accelerometer data conform to those shown in
the doc.
TEST=make buildall tests

CQ-DEPEND=CL:299013
CQ-DEPEND=CL:300510

Change-Id: Ib566f8eb2818c0694bf673af244aedbc91eac80f
Signed-off-by: Aseda Aboagye <aaboagye@google.com>