Cr50: Add support for flash write & erase
This adds flash support for the SoC.

BUG=chrome-os-partner:44745
BRANCH=none
TEST=manual

Pick an unused section of flash and use the flasherase and
flashwrite commands to test it. For example:

bank 0:

  md 0x7b000 20
  md 0x7b800 20
  flasherase 0x3b000 0x1000
  md 0x7b000 20
  md 0x7b800 20
  flashwrite 0x3b000 0x1000
  md 0x7b000 20
  md 0x7b800 20

bank 1:

  md 0xbb000 20
  md 0xbb800 20
  flasherase 0x7b000 0x1000
  md 0xbb000 20
  md 0xbb800 20
  flashwrite 0x7b000 0x1000
  md 0xbb000 20
  md 0xbb800 20

Note that the "md" command uses the absolute address, while the
flash commands use the offset address within the flash memory.

Change-Id: I956e813871949faed8d85ad9e46bdc64dee1a9e9
Signed-off-by: Bill Richardson <wfrichar@chromium.org>