Update patch set 1
Patch Set 1: Code-Review-1

This works in our case, but won't work all the time. If the source happens to send a message just after the sysjump, we won't respond to it (typically have 30ms to respond) and it will cause the source to send a hard reset. I think the better solution is to leave the soft reset before hard reset, but also implement chrome-os-partner:45133 where we wait for the ACCEPT/REJECT. After the ACCEPT/REJECT we can disable the TCPC with tcpm_set_rx_enable(0). This is also much more robust because it guarantees we wait long enough for the other side to get the soft reset instead of the blind delay we have now.

Patch-set: 1
Label: Code-Review=-1