driver/kionix_accel: Add 10ms sleep to init.
On certain platforms(glados) where we enable the sensor when
transitioning to S0, it appears that when we try to initialize the
sensor immediately, the sensor isn't quite ready yet and we fail.

The KX022 datasheet says that a typical power on time is 10ms.  This
commit simply adds a small delay at the beginning of init.

BUG=chrome-os-partner:43494
BRANCH=None
TEST=Flash GLaDOS EC and verify that upon boot, the sensor initializes
successfully.
TEST=make buildall tests

Change-Id: I53c79d69a3ae931443bcbc15101fe7db129df6f3
Signed-off-by: Aseda Aboagye <aaboagye@google.com>