Update patch set 2
Patch Set 2:

Luigi tested this on an MP Snow and it worked as intended. By holding a magnet over the lid switch he was able to simulate the lid being closed and press the power button to ensure it did not power on the system.

tl;dr version of the bug - The issue boils down to some TPM state being screwed up if the machine is turned on and then shut off within a short period of time. This can be caused by by stacking machines on top of another, squishing them, etc. So this patch works around the issue by preventing these accidental power button presses.

Patch-set: 2