Update patch set 1
Patch Set 1: Code-Review+2

Not sure if ectool console works (it prints error 74 twice, then exits) but we tested this before/after, and everything worked as expected.

We used a magnet to simulate lid closure.  With the lid "closed", before the change a power button push turns the device on.  After the change, it does not.  With the lid open, the power button works normally, before and after the change.

Patch-set: 1
Label: Code-Review=+2