Update patch set 2
Patch Set 2: Verified+1

As it currently stands, the lid angle reported is like that of samus. However, the thing I don't understand is that the angle seems *wrong*. Looking at the lid angle output, it seems that it's providing the complementary angle of the lid angle instead of the lid angle itself.

Is this expected?

Patch-set: 2
Label: Verified=+1