mkbp_event: prevent power connector insertion related events to wakeup AP.
It may use MKBP event to send PD power change event to AP by interrupt.
According to the spec, For power connector insertion, it does not allow to
wake up AP in S3, but, internal keyboard, lid ... does. So, prevent interrupt
AP in non-S0 state if the event is defined in NON_WAKEUP_EVENT_MASK.

BRANCH=none
BUG=chrome-os-partner:45127
TEST=manual
Plug PD power adapter to oak, if system is in S3/S5 and it should
starts charging, but should not wake up system/AP at all.

Change-Id: I2f86697d5d3bd24d7de840e21064b91e8841f0eb
Signed-off-by: Ben Lok <ben.lok@mediatek.com>