Update patch set 1
Patch Set 1:

Yeah. The board-specific build.mk #defines CHIP, which determines the chip-specific build.mk needed to #define CORE. Those are needed to run the preprocessor so that all the CONFIG_ definitions present in the compilation can be imported into Make as makefile variables.

Then we include the board-, chip-, and core-specific build.mk files again so that all the "FOO-$(CONFIG_BAR)+= ..." terms are valid.

It might be cleaner to put the CHIP definitions in a separate .mk file. OTOH, there's something to be said for having only one board-specific .mk file. YMMV.

Patch-set: 1