config_option_check.py: Ignore hidden files.
When this script is searching through CONFIG_* options used in the repo,
it should ignore hidden files.  I uncovered this by accidentally having
modified a file and the presubmit hook failed.  This is because emacs
auto-saves files with the format ".#filename.ext".  Now, the script will
ignore hidden files.

BUG=None
BRANCH=None
TEST=Use emacs to modify a c file but not save it. Run the script and
verify that no errors occur.
TEST=make -j buildall tests

Change-Id: Ie09cfcf66f2e416aeed99847945cb5168b457d1a
Signed-off-by: Aseda Aboagye <aaboagye@google.com>