Update patch set 4
Patch Set 4:

Regarding the question about PD_FLAGS_VCONN_ON, it wouldn't be an issue with a charge-through accessory or any device. We only check that flag in the VCONN swap states, and we can only get to those states once we have a contract (from the READY state), and we can only negotiate a contract once VBUS is applied. So, vconn will be on long before the flag is checked.

Patch-set: 4