Update patch set 1
Patch Set 1:

> I was told that if we use the 0.1c between 5 and 10C
> then it would halve the rate between 10 and 50C.

I don't fully understand this part.
Is that battery requirement (different from what is written in the bug ) ?

> So, the thought was to leave it at charging only between 10 and 50.  
> WDYT?

Whatever solution is the same for me,
but I have implemented what was supposed to be the final decision in the issue tracker, it would be nice to update it.

Patch-set: 1